merge_helpers.sh: Allow restoring upstream files deleted downstream

+++ merge_helpers.sh
@@ -67,7 +67,9 @@ restore_upstream() {
     local path="$(dirname "$1")"
     local sc_f="tmp_sc_$f"
     local upstream_f="upstream_$f"
-    mv "$path/$f" "$path/$sc_f"
+    if [ -e "$path/$f" ]; then
+        mv "$path/$f" "$path/$sc_f"
+    fi
     if [ -e "$path/$upstream_f" ]; then
         mv "$path/$upstream_f" "$path/$f"
     fi
